shared text resolver
